requirements.txt
plotly==5.14.0
inflection==0.5.1
pandas==1.4.2
streamlit==1.22.0
numpy==1.22.4
folium==0.14.0
haversine==2.8.0
Pillow==9.5.0